Create new file for map widget
