+++ lib/main.dart
@@ -1,8 +1,6 @@
-import "dart:async";
-
 import 'package:flutter/material.dart';
-import 'package:google_maps_flutter/google_maps_flutter.dart';
 
+import "/ui_layer/map.dart";
 import "/ui_layer/authentication.dart";
 import "/data_layer/database.dart";
 
@@ -50,39 +48,12 @@ class _MyHomePageState extends State<MyHomePage> {
           child: Column(
             children: <Widget>[
               GoogleSignInButton(),
-              //MapView(),
+              SizedBox(
+                height: 600,
+                child: InteractiveMap(),
+              )
             ],
           )
         ));
-  }
-}
-
-class MapView extends StatefulWidget {
-  const MapView({super.key});
-
-  @override
-  State<MapView> createState() => MapViewState();
-}
-
-class MapViewState extends State<MapView> {
-  final Completer<GoogleMapController> _controller =
-      Completer<GoogleMapController>();
-
-  static const CameraPosition _kGooglePlex = CameraPosition(
-    target: LatLng(50.79812, -1.09971),
-    zoom: 16.5,
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      body: GoogleMap(
-        mapType: MapType.normal,
-        initialCameraPosition: _kGooglePlex,
-        onMapCreated: (GoogleMapController controller) {
-          _controller.complete(controller);
-        },
-      ),
-    );
   }
 }
