+++ lib/ui_layer/map.dart
@@ -1,0 +1,34 @@
+import 'dart:async';
+import 'package:flutter/material.dart';
+import 'package:google_maps_flutter/google_maps_flutter.dart';
+
+class InteractiveMap extends StatefulWidget {
+  const InteractiveMap({super.key});
+
+  @override
+  State<InteractiveMap> createState() => _InteractiveMap();
+}
+
+class _InteractiveMap extends State<InteractiveMap> {
+  final Completer<GoogleMapController> _controller =
+      Completer<GoogleMapController>();
+
+  static const CameraPosition _kGooglePlex = CameraPosition(
+    target: LatLng(50.79812, -1.09971),
+    zoom: 16.5,
+  );
+
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      body: GoogleMap(
+        mapType: MapType.normal,
+        initialCameraPosition: _kGooglePlex,
+        onMapCreated: (GoogleMapController controller) {
+          _controller.complete(controller);
+        },
+      ),
+    );
+  }
+}
+
